samples/python/agents/a2a_telemetry/pyproject.toml
[project]
name = "a2a-telemetry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.6",
    "google-adk>=1.0.0",
    "opentelemetry-api>=1.33.1",
    "opentelemetry-exporter-jaeger>=1.11.1",
    "opentelemetry-exporter-otlp>=1.33.1",
    "opentelemetry-instrumentation-requests>=0.54b1",
    "opentelemetry-instrumentation-starlette>=0.54b1",
    "opentelemetry-sdk>=1.33.1",
]
